requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
defusedxml==0.7.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
Markdown==3.9
narwhals==2.19.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pydeck==0.9.2
pyparsing==3.3.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
youtube-transcript-api==1.2.4
